# swagger.yaml
openapi: 3.0.3
info:
  title: Trade-Back API
  description: |
    High-Performance Trading Backend API
    
    This API provides real-time market data streaming, technical indicators, 
    and historical data management for cryptocurrency trading.
    
    ## Features
    - Real-time price streaming via WebSocket (250-450+ msg/sec)
    - Historical data backfilling and management
    - Technical indicators (Enigma with Fibonacci levels)
    - Market watchlist management with auto-subscription
    - High-frequency data updates with binary protocol option
    - Session-based authentication for personalized features
    - Automatic symbol loading from Binance on startup
    
    ## WebSocket Protocol
    The WebSocket endpoint supports both JSON and binary protocols for optimal performance.
    Binary protocol reduces message size by 72% and parsing time by 16x.
    
    ## Market Watch Auto-Subscription
    When a symbol is added to the market watch list:
    1. All WebSocket connections with the same session token auto-subscribe
    2. Real-time prices start streaming immediately
    3. Only subscribed symbols receive price updates (efficient bandwidth usage)
    
    ## Rate Limits
    - REST API: 100 requests per minute per IP
    - WebSocket: No rate limit for subscriptions
    - Historical backfill: 10 requests per second (Binance limits)
  version: 1.0.0
  contact:
    name: Trading Team
    email: support@trade-back.com
  license:
    name: MIT
...
servers:
...
paths:
...
  /symbols:
    get:
      tags: [Market Data]
      summary: Get trading symbols
      description: Retrieve a list of available trading symbols with optional filtering
      operationId: getSymbols
      parameters:
        - name: exchange
          in: query
          description: Filter by exchange
          schema:
            type: string
            example: BINANCE
        - name: type
          in: query
          description: Filter by symbol type
          schema:
            type: string
            enum: [spot, futures]
            example: spot
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: List of symbols
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbols:
                    type: array
                    items:
                      $ref: '#/components/schemas/Symbol'
                  count:
                    type: integer
                    example: 350
...
  /historical/info:
    get:
      tags: [Historical]
      summary: Get historical API info
      description: Get information about the historical data API and available features
      operationId: getHistoricalInfo
      parameters:
        - name: symbol
          in: query
          description: Get info for specific symbol
          schema:
            type: string
            example: BTCUSDT
      responses:
        '200':
          description: Historical API information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoricalInfo'
  /historical/backfill/symbol:
    post:
      tags: [Historical]
      summary: Backfill single symbol
      description: |
        Load historical market data for a specific trading symbol.
        This operation runs asynchronously and returns a task ID for tracking.
        
        The backfill will:
        1. Fetch historical klines from Binance
        2. Store them in InfluxDB
        3. Update cache with latest data
        4. Calculate ATH/ATL if not exists
      operationId: backfillSymbol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackfillSymbolRequest'
            example:
              symbol: BTCUSDT
              interval: 1d
              days: 365
      responses:
        '202':
          description: Backfill task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackfillResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /historical/backfill/all:
    post:
      tags: [Historical]
      summary: Backfill all symbols
      description: |
        Load historical market data for all active trading symbols.
        This operation runs asynchronously and returns a task ID for tracking.
        
        Warning: This can take a long time and consume significant resources.
      operationId: backfillAll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackfillAllRequest'
            example:
              interval: 1d
              days: 730
      responses:
        '202':
          description: Backfill task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackfillResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /historical/ath-atl:
    post:
      tags: [Historical]
      summary: Calculate ATH/ATL
      description: |
        Calculate all-time high and all-time low values from historical data.
        If historical data is not available, it will first load 2 years of daily data.
      operationId: calculateATHATL
      responses:
        '202':
          description: ATH/ATL calculation started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
                  message:
                    type: string
                    example: ATH/ATL calculation started
                  task_id:
                    type: string
                    example: task_lqr8n3k4_g7h8i9
                  start_time:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalError'
  /historical/status/{taskId}:
    get:
      tags: [Historical]
      summary: Get task status
      description: Check the status of a historical data backfill task
      operationId: getTaskStatus
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID returned from backfill operation
          schema:
            type: string
            example: task_lqr8n3k2_a3b4c5
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Session token for authenticated endpoints
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid request parameters
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Symbol not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
...
    Symbol:
      type: object
      properties:
        id:
          type: integer
          example: 1
        symbol:
          type: string
          example: BTCUSDT
        exchange:
          type: string
          example: BINANCE
        full_name:
          type: string
          example: Bitcoin/Tether
        instrument_type:
          type: string
          example: SPOT
        base_currency:
          type: string
          example: BTC
        quote_currency:
          type: string
          example: USDT
        is_active:
          type: boolean
          example: true
        min_price_increment:
          type: number
          format: double
          example: 0.01
        min_quantity_increment:
          type: number
          format: double
          example: 0.00001
...
    BackfillSymbolRequest:
      type: object
      required: [symbol, interval, days]
      properties:
        symbol:
          type: string
          pattern: '^[A-Z]+$'
        interval:
          type: string
          enum: [1m, 3m, 5m, 15m, 30m, 1h, 2h, 4h, 6h, 8h, 12h, 1d, 3d, 1w, 1M]
        days:
          type: integer
          minimum: 1
          maximum: 1500
    BackfillAllRequest:
      type: object
      required: [interval, days]
      properties:
        interval:
          type: string
          enum: [1m, 3m, 5m, 15m, 30m, 1h, 2h, 4h, 6h, 8h, 12h, 1d, 3d, 1w, 1M]
        days:
          type: integer
          minimum: 1
          maximum: 1500
    BackfillResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        symbol:
          type: string
        symbols:
          type: array
          items:
            type: string
        interval:
          type: string
        days:
          type: integer
        start_time:
          type: string
          format: date-time
        task_id:
          type: string
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        error:
          type: string
    HistoricalInfo:
      type: object
      properties:
        message:
          type: string
        endpoints:
          type: object
          additionalProperties:
            type: string
        supported_intervals:
          type: array
          items:
            type: string
        max_days:
          type: integer
        symbol_info:
          type: object
          properties:
            symbol:
              type: string
            data_range:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                to:
                  type: string
                  format: date-time
...